Refactor Brave Ads timestamps from uint64_t to double and base::Time

+++ browser/extensions/api/brave_rewards_api.cc
@@ -1108,7 +1108,7 @@ BraveRewardsGetAdsAccountStatementFunction::Run() {
 
 void BraveRewardsGetAdsAccountStatementFunction::OnGetAdsAccountStatement(
     const bool success,
-    const int64_t next_payment_date,
+    const double next_payment_date,
     const int ads_received_this_month,
     const double earnings_this_month,
     const double earnings_last_month) {
