+++ browser/extensions/api/brave_rewards_api.h
@@ -400,7 +400,7 @@ class BraveRewardsGetAdsAccountStatementFunction : public ExtensionFunction {
 
  private:
   void OnGetAdsAccountStatement(const bool success,
-                                const int64_t next_payment_date,
+                                const double next_payment_date,
                                 const int ads_received_this_month,
                                 const double earnings_this_month,
                                 const double earnings_last_month);
